Create a CI test for the small_f90 mechanism in ci-tests/small_f90
In the ci-tests/small_f90 folder, we have made symbolic links to all
of the necessary files to build and run the small_f90 mechanism with
KPP.  The sequence to run the CI test is now

$ kpp small_f90.kpp
$ make
$ ./small_f90.exe

We have also made a symbolic link called Makefile that points to the
Makefile_small_f90 for convenience.

We have also updated the .gitignore files in ci-tests/small_f90
and in the examples folder so that Git will not store any of the
objects, modules, or executables.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>