Add script to run CI tests manually, for testing & debugging
We have added ./ci-pipelines/ci-manual-testing-script.sh, which allows
you to run all of the CI tests manually. It is similar to the
ci-testing-script.sh but is modified for the proper directory paths
(since ci-testing-script.sh is run within the Docker container).

Also changed the permission of ci-testing-script.sh from chmod 664
to chmod 775.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>