Pass ICNTRL to the solver in rosenbrock.f90 and rosenbrock_tlm.f90
int/rosenbrock.f90
int/rosenbrock_tlm.f90
- Now pass ICNTRL (i.e. KPP runtime options) from the top-level routine
  down to FunTemplate and JacTemplate via intermedate functions.
- Call Update_Sun() and Update_RCONST() if ICNTRL(15) == 3
- Only call Update_RCONST() if ICNTRL(15) == 1
- SKip calling Update_Sun() and Update_RCONST() within the solver step
  if ICNTRL(15) == 0
- Trimmed trailing whitespace

NOTE: The default value of ICNTRL(15)=0 should be OK because
Update_Sun() and Update_RCONST() is called within the CI test driver
programs.  This will also be true of models such as GEOS-Chem, where
we update the rates outside of the solver steps.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>