Temporarily restore the EQUIVALENCE statements in KPP_ROOT_Global.f90
For GEOS-Chem, we had removed the EQUIVALENCE statements (between
C and VAR and C and FIX) because these were not thread-safe when
calling KPP from the OpenMP parallel environment.  But this causes
the CI tests to fail.  Temporarily restore these statements until we
can implement a toggle that will let the user choose whether or not
to include them.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>