Add modifcation for #MODEL gckpp: Add optional argument Aout
For GEOS-Chem running in GEOS, there is need of passing the Aout argument
back from subroutine Fun.  We have added a test in gen.c to add the
appropriate code if the rootFileName is "gckpp", which is the name of
KPP implemented into GEOS-Chem.  Otherwise the modification is not
made.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>