Add function KppVersionIsTooOld into src/kpp.c
Added a function to check the minKppVersion value against the
current KPP version.  It will return 0 if the version is OK,
but will return 1 if the version is too old.  The minimum version
is specified by #MIN_VERSION in the definition file.

Also added an if block in the main routine to call this routine
and exit if the version is too old.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>