Use strncmp to test if minKppVersion == none to avoid issues on Azure
We have changed the test in function KppVersionIsTooOld so that we
use the strncmp command.  This allows us to restrict the search to
4 characters (the length of "none").

We had found that unless we wrote the test this way, the C-I tests
would fail (running in the Azure dev pipelines environment).  However,
the tests would pass if you ran them manually.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>