Reduce MAX_EQN and MAX_SPECIES in order to build KPP on MacOS
Long story short: MacOS X has a hardcoded max value for stack memory at 65532 kb.
Therefore, the current settings of MAX_EQN and MAX_SPECIES are too large, and
cause KPP to run out of memory, generating the infamous "Segmentation fault 11" error.

We have added an #ifdef MACOS block to set lower values of MAX_EQN and MAX_SPECIES
when building KPP on MacOs.

Also, it is recommended to add "ulimit -s 65532" to your .bashrc or .bash_aliases file
on MacOS so that the max amount of stack memory will be available in your login
environment.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>