Remove most symbolic links from ci-tests folders
KPP can look for the relevant files to include in the models/, drv/,
int/, and examples/ folders, so there is no need to link to these
files.  In each folder in ci-tests/, only keep the KPP definition file.

Also updated the .ci-pipelines/ci-manual-cleanup-script.sh to remove
the KPP-generated Makefile in each subfolder in ci-tests.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>