Added Lucas Estrada, Lu Shen, Daniel Jacob to list of authors
As per request of Daniel Jacob, we have added Lucas Estrada, Lu Shen,
and Daniel Jacob to the list of KPP authors on the front page of
ReadTheDocs (and also in .zenodo.json, which will trigger new DOIs).

Now all of the authors of the KPP 3.0.0 paper (plus Adrian Sandu)
are listed authors on the ReadTheDocs site.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>