uploading the implicit12 solver files
I have designed a new solver that is faster than the default Rosenbrock solvers.  
This new solver is inspired by the IMEX solver. I developed it using the standard BE and Sdirk2a because an explicit method would not work for stiff problems. 
This new solver gave me a 5X speed-up when I used it with the ICON-ART model. The speed-up outweighed the minor loss in accuracy.    

There are two files added:
- implicit12.f90
- implicit12.def